@@ -7,7 +7,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>1.5.2.RELEASE</version>
+        <version>2.4.3</version>
     </parent>    
 
     <groupId>com.sastix.cms</groupId>
@@ -27,13 +27,13 @@
 	
     <properties>
         <maven.test.failure.ignore>true</maven.test.failure.ignore>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <hazelcast.version>3.6.2</hazelcast.version>
-        <spring.boot.version>1.5.2.RELEASE</spring.boot.version>
-        <spring.retry.version>1.1.2.RELEASE</spring.retry.version>
-        <logback.version>1.1.2</logback.version>
-        <activiti.version>5.20.0</activiti.version>
-        <pdfbox.version>2.0.0</pdfbox.version>
+        <spring.boot.version>2.4.3</spring.boot.version>
+        <spring.retry.version>1.3.1</spring.retry.version>
+        <logback.version>1.2.3</logback.version>
+        <activiti.version>7.1.0</activiti.version>
+        <pdfbox.version>2.0.22</pdfbox.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     </properties>
