@@ -22,11 +22,12 @@ import com.sastix.cms.common.client.ApiVersionClient;
 import com.sastix.cms.common.client.RetryRestTemplate;
 import com.sastix.cms.common.client.impl.ApiVersionClientImpl;
 import com.sastix.cms.common.dataobjects.VersionDTO;
-import org.junit.Assert;
-import org.junit.Before;
+
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.mockito.*;
-import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,15 +38,14 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 
-import static org.junit.Assert.assertNotNull;
-
-
-@RunWith(SpringJUnit4ClassRunner.class)
+@TestInstance(Lifecycle.PER_CLASS)
 @SpringBootTest(classes = {CmsClient.class, CmsClient.class,VersionAPITest.class, RestTemplateConfiguration.class, ApiClientConfig.class},
 properties = {
         // cr client properties
@@ -69,9 +69,8 @@ public class VersionAPITest {
     @Mock
     ApiVersionClientImpl apiVersionClient;
 
-    @Before
+    @BeforeAll
     public void init() throws IOException, URISyntaxException {
-        MockitoAnnotations.initMocks(this);
         //spy syntax
         //Mockito.doReturn(TEST_VERSION).when(apiVersionClient).getApiVersion();
         Mockito.when(apiVersionClient.getApiVersion()).thenReturn(TEST_VERSION);
@@ -80,9 +79,9 @@ public class VersionAPITest {
     @Test
     public void getApiVersion() {
 
-        assertNotNull("service should be resolved", cmsClient);
+        assertNotNull(cmsClient, "service should be resolved");
         VersionDTO apiVersion = cmsClient.getApiVersion();
-        Assert.assertEquals("DTO should be the same", apiVersion, TEST_VERSION);
+        assertEquals(apiVersion, TEST_VERSION, "DTO should be the same");
 
         LOG.info("DTO returned by api call : {} ", apiVersion);
     }
