@@ -154,7 +154,7 @@ public class CommonExceptionHandler implements ResponseErrorHandler {
     private Charset getCharset(ClientHttpResponse response) {
         HttpHeaders headers = response.getHeaders();
         MediaType contentType = headers.getContentType();
-        return contentType != null ? contentType.getCharSet() : null;
+        return contentType != null ? contentType.getCharset() : null;
     }
 
 
