@@ -20,14 +20,13 @@ package com.sastix.cms.server;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-@Configuration
 @ComponentScan("com.sastix.cms")
-@EnableAutoConfiguration
+@SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
 @EnableTransactionManagement
 public class CmsServer {
 
